refactor: reduce qlty smells 39→20 to achieve maintainability A-grade
Targeted fixes for all 11 function-level smells:
- email_jmap_adapter.py: SyncContext namedtuple replaces 7-param functions,
  consolidated return statements via _resolve_parent_mailbox/_validate_move_targets
- toon.ts: decomposed parseDelimitedRow into parseQuotedChar/parseUnquotedChar,
  extracted needsCsvQuoting to fix boolean-logic smell
- config-hook.mjs: extracted registerClaudeCliModels and logConfigSummary
- browser_automation_utils.py: LikePostsConfig dataclass replaces 7 params,
  extracted _click_like_button and _scroll_for_more helpers
- quality-hooks.mjs: extracted handleToolBefore/handleToolAfter from factory

File-complexity splits (11 new modules):
- toon-parser.ts, toon-utils.ts (from toon.ts)
- agent-mcp-tools.mjs (from agent-loader.mjs)
- observability-sqlite.mjs (from observability.mjs)
- exec-commands.ts, logger.ts, router.ts, start.ts (from simplex-bot)
- pageindex_helpers.py (from pageindex-generator.py)
- email_frontmatter_utils.py (from email-thread-reconstruction.py)
- manifest_collectors.py (from generate-manifest.py)

Structural improvements:
- simplex-bot dispatchEvent: 10-case switch → handler map (saves ~9 complexity)
- Ratcheted QLTY_SMELL_THRESHOLD from 109 to 22 (20 actual + 2 buffer)
