+++ .agents/AGENTS.md
@@ -85,7 +85,7 @@ Task IDs: `/new-task` or `claim-task-id.sh`. NEVER grep TODO.md for next ID.
 
 **Model tiers**: Use GitHub labels to set the model tier. The pulse reads these labels for tier routing, not `model:` in `TODO.md`. See `reference/task-taxonomy.md`. **Brief quality determines which model tier can execute** — never assign a tier without verifying the brief meets that tier's prerequisites:
 
-- `tier:simple`: Haiku — requires a brief with verbatim code blocks, explicit file paths, and copy-pasteable implementation. **Hard disqualifiers:** >2 files, skeleton code blocks, error/fallback logic to design, estimate >1h, >4 acceptance criteria, judgment keywords (see `reference/task-taxonomy.md` "Tier Assignment Validation"). Never assign without checking the disqualifier list. Issues tagged `tier:simple` also automatically receive GitHub's `good first issue` label (applied by `_apply_tier_label_replace` in `issue-sync-helper.sh`, t2099) so they surface via the built-in community filter and external discovery sites — never auto-removed on escalation, matching the protected-label policy.
+- `tier:simple`: Haiku — requires a brief with verbatim code blocks, explicit file paths, and copy-pasteable implementation. **Hard disqualifiers:** >2 files, skeleton code blocks, error/fallback logic to design, estimate >1h, >4 acceptance criteria, judgment keywords (see `reference/task-taxonomy.md` "Tier Assignment Validation"). Never assign without checking the disqualifier list.
 - `tier:standard`: Sonnet — standard implementation, bug fixes, refactors. Narrative briefs with file references are sufficient. Use when uncertain. This is the default tier.
 - `tier:thinking`: Opus — architecture, novel design with no existing pattern to follow, deep reasoning, security audits.
 - **Cascade dispatch**: The pulse may start at `tier:simple` and escalate through tiers if the worker fails, accumulating context at each level. See `reference/task-taxonomy.md` "Cascade Dispatch Model".
