+++ .agents/configs/complexity-thresholds.conf
@@ -120,13 +120,19 @@ FILE_SIZE_THRESHOLD=59
 # namerefs/associative arrays. Proper fix is to refactor those scripts.
 BASH32_COMPAT_THRESHOLD=72
 
-# Qlty maintainability smell count — multi-language counter (GH#18775, t2067).
-# Unlike the shell-only thresholds above, this covers Python, JS/TS/mjs, and
-# cyclomatic complexity via `qlty smells --all` SARIF output. Enforced by the
-# qlty-smell-threshold job in code-quality.yml; auto-ratcheted down by
-# ratchet-post-merge.yml when the count drops below threshold - 2.
-# Baseline: 111 (2026-04-14, GH#18775): actual count 109 + 2 buffer
-QLTY_SMELL_THRESHOLD=111
+# Qlty maintainability smell baseline (t2065, GH#18773). Seed value for
+# the `.github/workflows/qlty-regression.yml` gate. The gate itself uses
+# a base-vs-head diff (not this threshold directly), but anchoring the
+# ratchet here gives t2067 (ratchet-down discipline) a starting point.
+# Baseline: 109 (2026-04-14, measured against main with
+# `qlty smells --all --sarif | jq '.runs[0].results | length'`).
+# Ratcheted down to 22 (refactor/qlty-a-grade): 39→20 smells via function-level
+# fixes (SyncContext, consolidated returns, decomposed parseDelimitedRow, config-hook
+# substeps, LikePostsConfig dataclass, quality-hooks extraction) + file splits
+# (toon-parser.ts, toon-utils.ts, agent-mcp-tools.mjs, observability-sqlite.mjs,
+# exec-commands.ts, logger.ts, router.ts, start.ts, pageindex_helpers.py,
+# email_frontmatter_utils.py, manifest_collectors.py). 20 actual + 2 buffer.
+QLTY_SMELL_THRESHOLD=22
 
 # Qlty smell-count → grade mapping (t2066, GH#18774)
 #
