+++ .agents/scripts/lib/browser_automation_utils.py
@@ -10,10 +10,22 @@ eliminate duplication (27 similar lines, mass=149 per qlty smells).
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, Callable
 
 if TYPE_CHECKING:
     from playwright.sync_api import Page
+
+
+@dataclass
+class LikePostsConfig:
+    """Configuration for the post-liking session."""
+    max_likes: int
+    session_stats: dict
+    random_delay_fn: Callable
+    like_count_key: str = "likes_given"
+    processed_count_key: str | None = "posts_processed"
+    label: str = ""
 
 
 def playwright_login(
@@ -60,78 +72,65 @@ def playwright_login(
         return False
 
 
-def playwright_like_posts(
-    page: "Page",
-    max_likes: int,
-    session_stats: dict,
-    random_delay_fn,
-    like_count_key: str = "likes_given",
-    processed_count_key: str | None = "posts_processed",
-    label: str = "",
-) -> None:
+def _click_like_button(button, cfg, likes_given, label_suffix):
+    """Click a single like button and update stats.
+
+    Returns updated likes_given count, or the same count on error.
+    """
+    try:
+        button.scroll_into_view_if_needed()
+        cfg.random_delay_fn(1, 3)
+        button.click()
+        likes_given += 1
+        cfg.session_stats[cfg.like_count_key] += 1
+        if cfg.processed_count_key is not None:
+            cfg.session_stats[cfg.processed_count_key] += 1
+        print(f"Liked post {likes_given}/{cfg.max_likes}{label_suffix}")
+        cfg.random_delay_fn()
+    except Exception as e:
+        print(f"Error liking post: {e}")
+        cfg.session_stats["errors"] += 1
+    return likes_given
+
+
+def _scroll_for_more(page, cfg):
+    """Scroll the page to load more content."""
+    page.evaluate("window.scrollTo(0, document.body.scrollHeight)")
+    cfg.random_delay_fn(3, 6)
+
+
+def playwright_like_posts(page: "Page", cfg: LikePostsConfig) -> None:
     """Like posts on LinkedIn timeline using Playwright.
 
     Args:
         page: Playwright page object.
-        max_likes: Maximum number of posts to like.
-        session_stats: Mutable stats dict updated in-place.
-        random_delay_fn: Callable matching ``random_delay(min, max)`` signature.
-        like_count_key: Key in ``session_stats`` to increment for each like.
-        processed_count_key: Optional key in ``session_stats`` to increment for
-            each processed post.  Pass ``None`` to skip.
-        label: Optional suffix appended to log messages (e.g. " (LOCAL browser)").
+        cfg: LikePostsConfig with session parameters.
     """
-    label_suffix = f" {label.strip()}" if label.strip() else ""
-    print(f"👍 Starting to like posts{label_suffix} (max: {max_likes})...")
+    label_suffix = f" {cfg.label.strip()}" if cfg.label.strip() else ""
+    print(f"Starting to like posts{label_suffix} (max: {cfg.max_likes})...")
 
     likes_given = 0
     scroll_attempts = 0
     max_scrolls = 5
 
-    while likes_given < max_likes and scroll_attempts < max_scrolls:
-        # Find like buttons that haven't been clicked
+    while likes_given < cfg.max_likes and scroll_attempts < max_scrolls:
         like_buttons = page.query_selector_all(
             'button[aria-label*="Like"][aria-pressed="false"]'
         )
 
         if not like_buttons:
-            print("📜 Scrolling to load more posts...")
-            page.evaluate("window.scrollTo(0, document.body.scrollHeight)")
-            random_delay_fn(3, 6)
+            print("Scrolling to load more posts...")
+            _scroll_for_more(page, cfg)
             scroll_attempts += 1
             continue
 
-        # Like posts with random selection
-        for button in like_buttons[: min(3, max_likes - likes_given)]:
-            try:
-                # Scroll button into view
-                button.scroll_into_view_if_needed()
-                random_delay_fn(1, 3)
+        for button in like_buttons[: min(3, cfg.max_likes - likes_given)]:
+            likes_given = _click_like_button(button, cfg, likes_given, label_suffix)
+            if likes_given >= cfg.max_likes:
+                break
 
-                # Click like button
-                button.click()
-                likes_given += 1
-                session_stats[like_count_key] += 1
-                if processed_count_key is not None:
-                    session_stats[processed_count_key] += 1
-
-                print(f"👍 Liked post {likes_given}/{max_likes}{label_suffix}")
-
-                # Random delay between likes
-                random_delay_fn()
-
-                if likes_given >= max_likes:
-                    break
-
-            except Exception as e:
-                print(f"⚠️ Error liking post: {e}")
-                session_stats["errors"] += 1
-                continue
-
-        # Scroll for more content
-        if likes_given < max_likes:
-            page.evaluate("window.scrollTo(0, document.body.scrollHeight)")
-            random_delay_fn(3, 6)
+        if likes_given < cfg.max_likes:
+            _scroll_for_more(page, cfg)
             scroll_attempts += 1
 
-    print(f"✅ Completed{label_suffix} liking session: {likes_given} posts liked")
+    print(f"Completed{label_suffix} liking session: {likes_given} posts liked")
