+++ .agents/scripts/local-browser-automation.py
@@ -33,7 +33,7 @@ except ImportError:
 
 # Shared browser automation utilities
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), "lib"))
-from browser_automation_utils import playwright_login, playwright_like_posts  # noqa: E402
+from browser_automation_utils import playwright_login, playwright_like_posts, LikePostsConfig  # noqa: E402
 
 try:
     from selenium import webdriver
@@ -197,15 +197,15 @@ class LocalBrowserAutomation:
 
     def _playwright_like_posts(self, page: Page, max_likes: int):
         """Like posts on LinkedIn timeline using LOCAL Playwright"""
-        playwright_like_posts(
-            page,
-            max_likes,
-            self.session_stats,
-            self.random_delay,
+        cfg = LikePostsConfig(
+            max_likes=max_likes,
+            session_stats=self.session_stats,
+            random_delay_fn=self.random_delay,
             like_count_key="actions_performed",
             processed_count_key=None,
             label="LOCAL browser",
         )
+        playwright_like_posts(page, cfg)
 
     def _print_session_summary(self):
         """Print automation session summary"""
