+++ .agents/scripts/simplex-bot/src/commands.ts
@@ -10,28 +10,13 @@
 
 import { resolve } from "node:path";
 import type { CommandContext, CommandDefinition } from "./types";
-import { ApprovalManager, executeShellCommand } from "./approval";
 import pkg from "../package.json";
 
-// =============================================================================
-// Shared Approval Manager
-// =============================================================================
-
-/**
- * Singleton approval manager — shared across all command handlers.
- * Initialised with defaults; the bot can reconfigure via setApprovalManager().
- */
-let approvalManager = new ApprovalManager();
-
-/** Replace the approval manager (called by SimplexAdapter after config is loaded) */
-export function setApprovalManager(manager: ApprovalManager): void {
-  approvalManager = manager;
-}
-
-/** Get the current approval manager (for shutdown cleanup) */
-export function getApprovalManager(): ApprovalManager {
-  return approvalManager;
-}
+// Re-export exec approval manager from extracted module
+export { setApprovalManager, getApprovalManager } from "./exec-commands";
+import {
+  runCommand, approveCommand, rejectCommand, pendingCommand,
+} from "./exec-commands";
 
 // =============================================================================
 // Built-in Commands
@@ -155,176 +140,6 @@ const taskCommand: CommandDefinition = {
       description +
       '"\n\n(Task creation not yet connected to TODO.md pipeline. This is a scaffold.)'
     );
-  },
-};
-
-/**
- * Execute a command remotely (DM only).
- *
- * Three-tier classification:
- *   - allowlist: executes immediately (e.g., "aidevops status")
- *   - approval-required: sends approval request, waits for /approve <id>
- *   - blocklist: always rejected (e.g., "rm -rf")
- */
-const runCommand: CommandDefinition = {
-  name: "run",
-  description: "Execute a command remotely (approval required for unsafe commands)",
-  groupEnabled: false,
-  dmEnabled: true,
-  handler: async (ctx: CommandContext): Promise<string> => {
-    const command = ctx.args.join(" ");
-    if (!command) {
-      return "Usage: /run [command]\nExample: /run aidevops status";
-    }
-
-    const contactId = ctx.contact?.contactId;
-    const contactName = ctx.contact?.localDisplayName ?? "unknown";
-    if (contactId === undefined) {
-      return "Cannot execute commands: no contact identity available.";
-    }
-
-    const classification = approvalManager.classify(command);
-
-    switch (classification) {
-      case "blocked":
-        return (
-          "BLOCKED: This command matches a blocked pattern and cannot be executed.\n" +
-          "Command: " + command
-        );
-
-      case "allowed": {
-        // Execute immediately — command is on the allowlist
-        const result = await executeShellCommand(command);
-        const lines = ["Executed (allowlisted):", ""];
-        if (result.stdout) {
-          lines.push(result.stdout);
-        }
-        if (result.stderr) {
-          lines.push("stderr: " + result.stderr);
-        }
-        lines.push("", "Exit code: " + result.exitCode);
-        return lines.join("\n");
-      }
-
-      case "approval-required": {
-        // Create pending approval request
-        const request = approvalManager.createRequest(
-          command,
-          contactId,
-          contactName,
-          ctx.reply,
-        );
-
-        return [
-          "Approval required for command execution.",
-          "",
-          "Command: " + command,
-          "Request ID: " + request.id,
-          "Timeout: " + approvalManager.formatTimeout(),
-          "",
-          "To approve:  /approve " + request.id,
-          "To reject:   /reject " + request.id,
-          "To list all: /pending",
-        ].join("\n");
-      }
-    }
-  },
-};
-
-/** Approve a pending command execution */
-const approveCommand: CommandDefinition = {
-  name: "approve",
-  description: "Approve a pending command execution",
-  groupEnabled: false,
-  dmEnabled: true,
-  handler: async (ctx: CommandContext): Promise<string> => {
-    const requestId = ctx.args[0];
-    if (!requestId) {
-      return "Usage: /approve [request-id]\nExample: /approve a3f7";
-    }
-
-    const contactId = ctx.contact?.contactId;
-    if (contactId === undefined) {
-      return "Cannot approve: no contact identity available.";
-    }
-
-    const request = approvalManager.approve(requestId, contactId);
-    if (!request) {
-      // Check if the request exists but is in a non-pending state
-      const existing = approvalManager.getRequest(requestId);
-      if (existing) {
-        return "Request [" + requestId + "] is no longer pending (state: " + existing.state + ").";
-      }
-      return "No pending request found with ID: " + requestId;
-    }
-
-    // Execute the approved command
-    await ctx.reply("Approved. Executing: " + request.command);
-
-    const result = await executeShellCommand(request.command);
-    const lines = ["Result for [" + requestId + "]:", ""];
-    if (result.stdout) {
-      lines.push(result.stdout);
-    }
-    if (result.stderr) {
-      lines.push("stderr: " + result.stderr);
-    }
-    lines.push("", "Exit code: " + result.exitCode);
-    return lines.join("\n");
-  },
-};
-
-/**
- * Reject a pending command execution.
- *
- * Authorization note: /reject intentionally has no identity check (unlike /approve
- * which validates contactId). This allows any DM contact to reject suspicious
- * commands — a safety-first design. To restrict rejection to the original
- * requester, pass contactId to approvalManager.reject() (see approval.ts).
- */
-const rejectCommand: CommandDefinition = {
-  name: "reject",
-  description: "Reject a pending command execution",
-  groupEnabled: false,
-  dmEnabled: true,
-  handler: async (ctx: CommandContext): Promise<string> => {
-    const requestId = ctx.args[0];
-    if (!requestId) {
-      return "Usage: /reject [request-id]\nExample: /reject a3f7";
-    }
-
-    const request = approvalManager.reject(requestId);
-    if (!request) {
-      return "No pending request found with ID: " + requestId;
-    }
-
-    return "Rejected command [" + requestId + "]: " + request.command;
-  },
-};
-
-/** List pending approval requests */
-const pendingCommand: CommandDefinition = {
-  name: "pending",
-  description: "List pending command approval requests",
-  groupEnabled: false,
-  dmEnabled: true,
-  handler: async (ctx: CommandContext): Promise<string> => {
-    const contactId = ctx.contact?.contactId;
-    const requests = approvalManager.listPending(contactId);
-
-    if (requests.length === 0) {
-      return "No pending approval requests.";
-    }
-
-    const lines = ["Pending approval requests:", ""];
-    for (const req of requests) {
-      const ageSeconds = Math.round((Date.now() - req.createdAt) / 1000);
-      lines.push(
-        "[" + req.id + "] " + req.command + " (" + ageSeconds + "s ago)",
-      );
-    }
-    lines.push("", "Use /approve <id> or /reject <id> to respond.");
-    return lines.join("\n");
   },
 };
