+++ TODO.md
@@ -621,7 +621,6 @@ t165,Provider-agnostic task claiming via TODO.md,marcusquinn,orchestration archi
 
 ## Done
 
-- [x] t2099 feat(issue-sync): auto-apply good first issue label to tier:simple issues — extend _apply_tier_label_replace in issue-sync-helper.sh so whenever tier:simple is applied, the good first issue label is also ensured via gh_create_label and added. Only added on tier:simple, never auto-removed on escalation (label is already in _is_protected_label and may reflect human intent). Surfaces prescriptive single-file pattern-follow tasks to GitHub standard community discovery filters. Adds test assertion and one-line AGENTS.md note. Backfill open tier:simple issues across pulse repos after merge. #feat #community #auto-dispatch #interactive ~45m tier:simple ref:GH#19007 logged:2026-04-14 -> [todo/tasks/t2099-brief.md] pr:#19008 completed:2026-04-14
 - [x] t2054 fix(post-merge-review-scanner): GraphQL reviewThreads resolution filter + diffHunk context + refresh backfill — issue #18621 (now closed) showed every CodeRabbit finding was already resolved by CodeRabbit itself, yet the scanner created a spurious followup. Replace REST /comments with GraphQL `reviewThreads(first: 100)` and filter `isResolved` / `isOutdated` — canonical bot-agnostic signals that work for CodeRabbit, Gemini, claude-review, gpt-review uniformly. Add diffHunk tail (last 12 lines) as a ```diff fence under each comment so workers see flagged code inline. Drop CodeRabbit `Actionable comments posted: N` metadata summaries. Add `refresh` / `refresh-dry-run` subcommands that rewrite open review-followup issue bodies and close issues whose source PR has zero unresolved threads — self-heals stale backlog from pre-PR#18607 scanner runs and future template changes. Add test-post-merge-review-scanner.sh (27 stub-based assertions). #bug #scanner #auto-dispatch #interactive ~3h tier:standard ref:GH#18734 logged:2026-04-13 -> [todo/tasks/t2054-brief.md] pr:#18736 completed:2026-04-13
 - [x] t2063 fix: inline brief into issue body so workers receive full context on first dispatch — root cause of Sonnet workers failing to solve issues on first pass. Two recent examples (GH#18745 t2059, GH#18746 t2060) had 476-char pointer bodies despite 10-12 KB brief files. Caused by ordering: claim-task-id.sh creates the issue before TODO.md has the ref, so delegation to issue-sync push fails and the bare fallback uses --description verbatim. Then _enrich_update_issue preserves the stub because it lacks the "Synced from TODO.md" sentinel. Fix: (A) claim-task-id _compose_issue_body inlines brief when one exists, appends sentinel; (B) _enrich_update_issue uses brief-file presence as authoritative signal instead of sentinel; (C) dispatch-time guard force-enriches if brief exists but body lacks Task Brief marker; plus case-insensitive heading matcher and template heading-lock. #bug #auto-dispatch #interactive ~3.5h tier:standard ref:GH#18771 logged:2026-04-14 -> [todo/tasks/t2063-brief.md] pr:#18782 completed:2026-04-14
 - [x] t2046 parent-task lifecycle hardening — fail-closed dispatch guard + parent-task PR keyword guard. Two systemic fixes from the t2010/GH#18458 incident root-cause analysis. Deliverable A: switch is_assigned() to fail-closed on internal errors (jq, gh API, helper failures) — emit GUARD_UNCERTAIN signal mirroring PARENT_TASK_BLOCKED pattern. Architectural belt-and-braces beyond the GH#18537 specific-bug fix. Deliverable B: prevent the Resolves #NNN auto-close trap that bit PR #18581 — three layered defenses: brief-template + AGENTS.md doc paragraph, client-side parent-task-keyword-guard.sh wired into commit-and-pr, GitHub Actions belt-and-braces workflow. Plan: todo/plans/parent-task-incident-hardening.md (root cause + design). #bug #pulse #refactor #auto-dispatch ~5.5h tier:standard ref:GH#18599 logged:2026-04-13 -> [todo/tasks/t2046-brief.md] -> [todo/plans/parent-task-incident-hardening.md] pr:#18663 completed:2026-04-13
